refactor: pluralize users module
