+++ src/users/users.module.ts
@@ -1,7 +1,7 @@
 import { MiddlewareConsumer, Module, NestModule, RequestMethod } from '@nestjs/common';
-import { UserService } from './user.service';
+import { UsersService } from './users.service';
 import { SequelizeModule } from '@nestjs/sequelize';
-import { UserController } from './user.controller';
+import { UsersController } from './users.controller';
 import { AuthModule } from '../auth/auth.module';
 import { User } from './user.model';
 import { UserMiddleware } from './middleware/user.middleware';
@@ -11,11 +11,11 @@ import { UserMiddleware } from './middleware/user.middleware';
     SequelizeModule.forFeature([User]),
     AuthModule
   ],
-  providers: [UserService],
-  controllers: [UserController],
+  providers: [UsersService],
+  controllers: [UsersController],
   exports: [SequelizeModule],
 })
-export class UserModule implements NestModule {
+export class UsersModule implements NestModule {
   configure(consumer: MiddlewareConsumer) {
     consumer.apply(UserMiddleware).forRoutes(
       { path: 'users/:userId', method: RequestMethod.GET }
